+++ CHANGELOG.md
@@ -1,3 +1,28 @@
+# [1.4.0-next.2](https://github.com/Print-one/print-one-js/compare/v1.4.0-next.1...v1.4.0-next.2) (2024-08-27)
+
+
+### Features
+
+* :sparkles: added coupon-code-used webhook event handler ([ba28867](https://github.com/Print-one/print-one-js/commit/ba2886762e6368a5254a8b1081f0965e5f426d50))
+* :sparkles: rename x-printone webhook headers to x-webhook ([4b3b818](https://github.com/Print-one/print-one-js/commit/4b3b8189f1e023279e878d53e6467166edbb80f7))
+
+# [1.4.0-next.1](https://github.com/Print-one/print-one-js/compare/v1.3.0...v1.4.0-next.1) (2024-08-07)
+
+
+### Bug Fixes
+
+* :rotating_light: fix linting issues ([7bf02e4](https://github.com/Print-one/print-one-js/commit/7bf02e458c833554fc3e661497a7233f7a9be0ea))
+* :white_check_mark: skip coupon code usage test, if order is not billable ([4a9ff10](https://github.com/Print-one/print-one-js/commit/4a9ff103fd166af87c7010ab585f5c91af9b79cb))
+* typescript error on install ([20d3f0f](https://github.com/Print-one/print-one-js/commit/20d3f0fc8abadde35381dda455f1ae1d8775bc94))
+
+
+### Features
+
+* :sparkles: Added coupons ([604c374](https://github.com/Print-one/print-one-js/commit/604c3747a6587349f0040fba1838ef4a7a77e19d))
+* :sparkles: webhook request validation logic ([ae9bd7c](https://github.com/Print-one/print-one-js/commit/ae9bd7c9c67b6516bcd0d47616162bc93c01282b))
+* add batch_status_update webhook support ([696a1dc](https://github.com/Print-one/print-one-js/commit/696a1dc2f0e4dc124ae43e0fcd654574c9c07509))
+* webhooks ([73a1678](https://github.com/Print-one/print-one-js/commit/73a1678a490618251a0a9b80a77931ad8b31e0f0))
+
 # [1.3.0](https://github.com/Print-one/print-one-js/compare/v1.2.1...v1.3.0) (2024-07-01)
 
 
@@ -21,6 +46,28 @@
 * added format to DTO ([64ec3ad](https://github.com/Print-one/print-one-js/commit/64ec3adc8b51dd697d0c1244e663353b087a17c1))
 * batch endpoints added ([#8](https://github.com/Print-one/print-one-js/issues/8)) ([0099217](https://github.com/Print-one/print-one-js/commit/009921704b0c7b75206341ef20b1f540c31e366c))
 * implement PR feedback ([8420046](https://github.com/Print-one/print-one-js/commit/8420046c3655a0a9480f192f3879fcf9524c06f4))
+
+# [1.3.0-next.3](https://github.com/Print-one/print-one-js/compare/v1.3.0-next.2...v1.3.0-next.3) (2024-07-02)
+
+
+### Features
+
+* add batch_status_update webhook support ([696a1dc](https://github.com/Print-one/print-one-js/commit/696a1dc2f0e4dc124ae43e0fcd654574c9c07509))
+
+# [1.3.0-next.2](https://github.com/Print-one/print-one-js/compare/v1.3.0-next.1...v1.3.0-next.2) (2024-07-01)
+
+
+### Bug Fixes
+
+* typescript error on install ([8fb29c4](https://github.com/Print-one/print-one-js/commit/8fb29c4f83831d12510e31475fb3a6fdc4ed352f))
+* typescript error on install ([20d3f0f](https://github.com/Print-one/print-one-js/commit/20d3f0fc8abadde35381dda455f1ae1d8775bc94))
+
+
+### Features
+
+* :sparkles: webhook request validation logic ([ae9bd7c](https://github.com/Print-one/print-one-js/commit/ae9bd7c9c67b6516bcd0d47616162bc93c01282b))
+* :sparkles: webhook request validation logic ([ce81792](https://github.com/Print-one/print-one-js/commit/ce81792e9b554ceaf0694857a9402b003bb80392))
+* webhooks ([73a1678](https://github.com/Print-one/print-one-js/commit/73a1678a490618251a0a9b80a77931ad8b31e0f0))
 
 # [1.3.0-next.1](https://github.com/Print-one/print-one-js/compare/v1.2.1...v1.3.0-next.1) (2024-05-22)